+++ adapters/google_cloud_storage.go
@@ -27,12 +27,12 @@ type GoogleConfig struct {
 	credentials option.ClientOption
 }
 
-func (gc *GoogleConfig) Validate(streamingMode bool) error {
+func (gc *GoogleConfig) Validate(streamMode bool) error {
 	if gc == nil {
 		return errors.New("Google config is required")
 	}
 	//batch mode works via google cloud storage
-	if !streamingMode && gc.Bucket == "" {
+	if !streamMode && gc.Bucket == "" {
 		return errors.New("Google cloud storage bucket(gcs_bucket) is required parameter")
 	}
 	if gc.Project == "" {