+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 
 ## [v1.2.13] - 2020-09-01
 ### Added
-- [18](https://github.com/ksensehq/eventnative/issues/18) Two destination modes: streaming and batch
+- [18](https://github.com/ksensehq/eventnative/issues/18) Two destination modes: stream and batch
 
 ## [v1.2.12] - 2020-09-01
 ### Fixed